+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## v0.2.1 Feature release
+* Added support for source base directory
+
+## v0.2.0 Feature release
+* Added support for base directory globs
+
 ## v0.1.0 Initial release
 * Basic file concatenation & flattening